refactor: extract InstanceConfig from Config singleton (Phase 1)
Create InstanceConfig class that mirrors all Config singleton fields
but is instantiable per-chain-instance. object Config extends
InstanceConfig for full backward compatibility — all existing code
continues to work unchanged.

This is the foundation for multi-instance runtime: each ChainInstance
will create its own InstanceConfig with different HOCON configuration,
database paths, ports, and chain parameters.

Changes:
- New InstanceConfig class with all Config fields (config, Network,
  Db, blockchains, supportedCapabilities, etc.)
- New InstanceConfigProvider trait for cake pattern integration
- Move NodeCacheConfig trait to package level for type accessibility
- object Config extends InstanceConfig(defaultConfig) as bridge
- Fix 3 call sites: AppCaches, Storages, LruCache/MapCache/StateStorage
