+++ src/main/scala/com/chipprbots/ethereum/db/components/Storages.scala
@@ -54,7 +54,7 @@ object Storages {
           pruningMode,
           nodeStorage,
           new LruCache[NodeHash, HeapEntry](
-            Config.InMemoryPruningNodeCacheConfig,
+            Config.inMemoryPruningNodeCacheConfig,
             Some(CachedReferenceCountedStorage.saveOnlyNotificationHandler(nodeStorage))
           )
         )
