+++ src/main/scala/com/chipprbots/ethereum/utils/Config.scala
@@ -26,110 +26,11 @@ import com.chipprbots.ethereum.utils.VmConfig.VmMode
 
 import ConfigUtils._
 
-object Config {
-
-  val config: TypesafeConfig = ConfigFactory.load().getConfig("fukuii")
-
-  val testmode: Boolean = config.getBoolean("testmode")
-
-  val clientId: String =
-    VersionInfo.nodeName(ConfigUtils.getOptionalValue(config, _.getString, "client-identity"))
-
-  val clientVersion: String = VersionInfo.nodeName()
-
-  val nodeKeyFile: String = config.getString("node-key-file")
-
-  val shutdownTimeout: Duration = config.getDuration("shutdown-timeout").toMillis.millis
-
-  val secureRandomAlgo: Option[String] =
-    if (config.hasPath("secure-random-algo")) Some(config.getString("secure-random-algo"))
-    else None
-
-  // Node capabilities - determined by what this version of Fukuii supports
-  // Per DevP2P spec and Geth implementation: advertise ALL supported protocol versions
-  // While ETH versions are backward compatible, advertising all versions ensures
-  // proper capability negotiation with peers that may only support older versions.
-  // Geth advertises: eth/66, eth/67, eth/68, snap/1
-  // We advertise: eth/65, eth/66, eth/67, eth/68, snap/1 to maximize compatibility
-  //
-  // Note: ETH63 and ETH64 are legacy protocols from 2016-2019 and are not actively
-  // advertised in the Hello message. However, they remain supported during the
-  // negotiation phase for backward compatibility with very old clients.
-  // ETH65+ are advertised as they are still commonly used in the ecosystem.
-  //
-  // Historical note: ETC64 protocol support was removed in favor of standard ETH protocols.
-  // The client now exclusively supports ETH63-68 and SNAP1, aligning with Ethereum specifications.
-  // See docs/validation/ETC64_REMOVAL_VALIDATION.md for details.
-  import com.chipprbots.ethereum.network.p2p.messages.Capability
-  val supportedCapabilities: List[Capability] = List(
-    Capability.ETH65,
-    Capability.ETH66,
-    Capability.ETH67,
-    Capability.ETH68,
-    Capability.SNAP1
-  )
-
-  val blockchains: BlockchainsConfig = BlockchainsConfig(config.getConfig("blockchains"))
-
-  object Network {
-    private val networkConfig = config.getConfig("network")
-
-    val automaticPortForwarding: Boolean = networkConfig.getBoolean("automatic-port-forwarding")
-
-    object Server {
-      private val serverConfig = networkConfig.getConfig("server-address")
-
-      val interface: String = serverConfig.getString("interface")
-      val port: Int = serverConfig.getInt("port")
-      val listenAddress = new InetSocketAddress(interface, port)
-    }
-
-    val peer: PeerConfiguration = new PeerConfiguration {
-      private val peerConfig = networkConfig.getConfig("peer")
-      private val blockchainConfig: BlockchainConfig = blockchains.blockchainConfig
-
-      val connectRetryDelay: FiniteDuration = peerConfig.getDuration("connect-retry-delay").toMillis.millis
-      val connectMaxRetries: Int = peerConfig.getInt("connect-max-retries")
-      val disconnectPoisonPillTimeout: FiniteDuration =
-        peerConfig.getDuration("disconnect-poison-pill-timeout").toMillis.millis
-      val waitForHelloTimeout: FiniteDuration = peerConfig.getDuration("wait-for-hello-timeout").toMillis.millis
-      val waitForStatusTimeout: FiniteDuration = peerConfig.getDuration("wait-for-status-timeout").toMillis.millis
-      val waitForChainCheckTimeout: FiniteDuration =
-        peerConfig.getDuration("wait-for-chain-check-timeout").toMillis.millis
-      val minOutgoingPeers: Int = peerConfig.getInt("min-outgoing-peers")
-      val maxOutgoingPeers: Int = peerConfig.getInt("max-outgoing-peers")
-      val maxIncomingPeers: Int = peerConfig.getInt("max-incoming-peers")
-      val maxPendingPeers: Int = peerConfig.getInt("max-pending-peers")
-      val pruneIncomingPeers: Int = peerConfig.getInt("prune-incoming-peers")
-      val minPruneAge: FiniteDuration = peerConfig.getDuration("min-prune-age").toMillis.millis
-      val networkId: Int = blockchainConfig.networkId
-      // Default to v5 (highest known p2p version supporting Snappy compression) if not explicitly set
-      val p2pVersion: Int = if (peerConfig.hasPath("p2p-version")) peerConfig.getInt("p2p-version") else 5
-
-      val rlpxConfiguration: RLPxConfiguration = new RLPxConfiguration {
-        val waitForHandshakeTimeout: FiniteDuration =
-          peerConfig.getDuration("wait-for-handshake-timeout").toMillis.millis
-        val waitForTcpAckTimeout: FiniteDuration = peerConfig.getDuration("wait-for-tcp-ack-timeout").toMillis.millis
-      }
-
-      val fastSyncHostConfiguration: FastSyncHostConfiguration = new FastSyncHostConfiguration {
-        val maxBlocksHeadersPerMessage: Int = peerConfig.getInt("max-blocks-headers-per-message")
-        val maxBlocksBodiesPerMessage: Int = peerConfig.getInt("max-blocks-bodies-per-message")
-        val maxReceiptsPerMessage: Int = peerConfig.getInt("max-receipts-per-message")
-        val maxMptComponentsPerMessage: Int = peerConfig.getInt("max-mpt-components-per-message")
-      }
-      override val updateNodesInitialDelay: FiniteDuration =
-        peerConfig.getDuration("update-nodes-initial-delay").toMillis.millis
-      override val updateNodesInterval: FiniteDuration = peerConfig.getDuration("update-nodes-interval").toMillis.millis
-
-      val shortBlacklistDuration: FiniteDuration = peerConfig.getDuration("short-blacklist-duration").toMillis.millis
-      val longBlacklistDuration: FiniteDuration = peerConfig.getDuration("long-blacklist-duration").toMillis.millis
-
-      val statSlotDuration: FiniteDuration = peerConfig.getDuration("stat-slot-duration").toMillis.millis
-      val statSlotCount: Int = peerConfig.getInt("stat-slot-count")
-    }
-
-  }
+/** Singleton Config for backward compatibility.
+  * All existing code that references `Config.xxx` continues to work unchanged.
+  * For multi-instance mode, create new `InstanceConfig` instances instead.
+  */
+object Config extends InstanceConfig(ConfigFactory.load().getConfig("fukuii"), "default") {
 
   case class SyncConfig(
       doFastSync: Boolean,
@@ -277,44 +178,8 @@ object Config {
     }
   }
 
-  object Db {
-
-    private val dbConfig = config.getConfig("db")
-    private val rocksDbConfig = dbConfig.getConfig("rocksdb")
-
-    val dataSource: String = dbConfig.getString("data-source")
-    val periodicConsistencyCheck: Boolean = dbConfig.getBoolean("periodic-consistency-check")
-
-    object RocksDb extends RocksDbConfig {
-      override val createIfMissing: Boolean = rocksDbConfig.getBoolean("create-if-missing")
-      override val paranoidChecks: Boolean = rocksDbConfig.getBoolean("paranoid-checks")
-      override val path: String = rocksDbConfig.getString("path")
-      override val maxThreads: Int = rocksDbConfig.getInt("max-threads")
-      override val maxOpenFiles: Int = rocksDbConfig.getInt("max-open-files")
-      override val verifyChecksums: Boolean = rocksDbConfig.getBoolean("verify-checksums")
-      override val levelCompaction: Boolean = rocksDbConfig.getBoolean("level-compaction-dynamic-level-bytes")
-      override val blockSize: Long = rocksDbConfig.getLong("block-size")
-      override val blockCacheSize: Long = rocksDbConfig.getLong("block-cache-size")
-    }
-
-  }
-
-  trait NodeCacheConfig {
-    val maxSize: Long
-    val maxHoldTime: FiniteDuration
-  }
-
-  object NodeCacheConfig extends NodeCacheConfig {
-    private val cacheConfig = config.getConfig("node-caching")
-    override val maxSize: Long = cacheConfig.getInt("max-size")
-    override val maxHoldTime: FiniteDuration = cacheConfig.getDuration("max-hold-time").toMillis.millis
-  }
-
-  object InMemoryPruningNodeCacheConfig extends NodeCacheConfig {
-    private val cacheConfig = config.getConfig("inmemory-pruning-node-caching")
-    override val maxSize: Long = cacheConfig.getInt("max-size")
-    override val maxHoldTime: FiniteDuration = cacheConfig.getDuration("max-hold-time").toMillis.millis
-  }
+  // SyncConfig remains here as it's a case class used as a type throughout the codebase.
+  // Db, Network, and cache configs are inherited from InstanceConfig.
 }
 
 case class AsyncConfig(askTimeout: Timeout)
